# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NQL DZ Transport Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [driver, customer, distributor]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create driver or distributor account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [admin]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [admin]
      summary: Reset user password
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/cards:
    get:
      operationId: getAdminCards
      tags: [admin]
      summary: Get all cards
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, rejected]
      responses:
        "200":
          description: Cards list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardsListResponse"
  /admin/cards/{id}/approve:
    post:
      operationId: approveCard
      tags: [admin]
      summary: Approve a pending card
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /admin/cards/{id}/reject:
    post:
      operationId: rejectCard
      tags: [admin]
      summary: Reject a pending card
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /admin/transactions:
    get:
      operationId: getAdminTransactions
      tags: [admin]
      summary: Get all transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: type
          in: query
          schema:
            type: string
            enum: [ride, topup]
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsListResponse"
  /admin/withdrawal-requests:
    get:
      operationId: getWithdrawalRequests
      tags: [admin]
      summary: Get withdrawal requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Withdrawal requests list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequestsListResponse"
  /admin/withdrawal-requests/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve withdrawal request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
  /admin/withdrawal-requests/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject withdrawal request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
  /admin/distributors/{id}/topup:
    post:
      operationId: topupDistributor
      tags: [admin]
      summary: Add balance to distributor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupDistributorRequest"
      responses:
        "200":
          description: Topup successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # DRIVER
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    ResetPasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, driver, customer, distributor]
        profileImage:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        balance:
          type: number
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - lastName
        - email
        - phone
        - role
        - status
        - createdAt
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required:
        - users
        - total
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [driver, distributor]
        licenseNumber:
          type: string
        profileImage:
          type: string
          nullable: true
      required:
        - name
        - lastName
        - email
        - phone
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive]
        licenseNumber:
          type: string
...
    Card:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [standard, student, employee, special_needs]
        cardNumber:
          type: string
        balance:
          type: number
        status:
          type: string
          enum: [pending, active, rejected]
        studentId:
          type: string
          nullable: true
        documentUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - cardNumber
        - balance
        - status
        - createdAt
    CardsListResponse:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        total:
          type: integer
      required:
        - cards
        - total
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        cardId:
          type: integer
        driverId:
          type: integer
          nullable: true
        distributorId:
          type: integer
          nullable: true
        amount:
          type: number
        platformFee:
          type: number
        driverEarning:
          type: number
        type:
          type: string
          enum: [ride, topup]
        cardType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - cardId
        - amount
        - platformFee
        - driverEarning
        - type
        - createdAt
    TransactionsListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
      required:
        - transactions
        - total
...
    WithdrawalRequest:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, ccp]
        phone:
          type: string
          nullable: true
        ccpAccount:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - driverId
        - amount
        - method
        - status
        - createdAt
    WithdrawalRequestsListResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/WithdrawalRequest"
        total:
          type: integer
      required:
        - requests
        - total
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalDrivers:
          type: integer
        totalCustomers:
          type: integer
        totalDistributors:
          type: integer
        totalTransactions:
          type: integer
        totalPlatformEarnings:
          type: number
        pendingCards:
          type: integer
        pendingWithdrawals:
          type: integer
        todayTransactions:
          type: integer
        todayEarnings:
          type: number
      required:
        - totalUsers
        - totalDrivers
        - totalCustomers
        - totalDistributors
        - totalTransactions
        - totalPlatformEarnings
        - pendingCards
        - pendingWithdrawals
        - todayTransactions
        - todayEarnings
    TopupDistributorRequest:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount
...