# lib/api-zod/src/generated/types/cardsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NQL DZ Transport Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Card } from "./card";

export interface CardsListResponse {
  cards: Card[];
  total: number;
}