# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NQL DZ Transport Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  name: string;
  lastName: string;
  email: string;
  phone: string;
  role: UserRole;
  profileImage?: string | null;
  licenseNumber?: string | null;
  balance?: number | null;
  status: UserStatus;
  createdAt: Date;
}